moved parent from BasePageDocument to PageDocument
