+++ src/Sulu/Bundle/ContentBundle/Document/PageDocument.php
@@ -18,4 +18,28 @@ use Sulu\Component\DocumentManager\Behavior\Path\AutoNameBehavior;
  */
 class PageDocument extends BasePageDocument implements AutoNameBehavior
 {
+    /**
+     * @var object
+     */
+    private $parent;
+
+    /**
+     * Return the parent document for this document.
+     *
+     * @return object
+     */
+    public function getParent()
+    {
+        return $this->parent;
+    }
+
+    /**
+     * Set the parent document for this document.
+     *
+     * @param object $parent
+     */
+    public function setParent($parent)
+    {
+        $this->parent = $parent;
+    }
 }
