+++ src/Sulu/Bundle/ContentBundle/Resources/config/form.xml
@@ -21,9 +21,6 @@
         </service>
 
         <service id="dtl_content.form.type.home" class="Sulu\Bundle\ContentBundle\Form\Type\HomeDocumentType">
-            <argument type="service" id="sulu.phpcr.session" />
-            <argument type="service" id="sulu_document_manager.document_manager" />
-
             <tag name="form.type" alias="home" />
         </service>
