+++ src/Sulu/Bundle/ContentBundle/Resources/config/serializer/Document.PageDocument.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <serializer>
     <class name="Sulu\Bundle\ContentBundle\Document\PageDocument">
+        <property name="parent" exclude="true"/>
     </class>
 </serializer>
